requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.7
pandas>=2.0.0
streamlit-aggrid>=0.3.4
sqlalchemy>=2.0.0
pytest>=7.4.0
pytest-mock>=3.11.1
python-dotenv>=1.0.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0
loguru>=0.7.0
pyjwt[crypto]>=2.8.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
streamlit-authenticator>=0.2.3
cryptography>=41.0.0
plotly>=5.17.0
reportlab>=4.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
psutil>=5.9.0
